Rock-Paper-Scissors Added UI
This is a copy of my Rock Paper Scissors project. I added an HTML and CSS file, and reworked some code. I had to copy this to a new repo b/c I am still learning how to branch and commit. In an effort not to lose my work, I am adding this to a new repo.